Refactor ViewWeaponDetailsFragment
