+++ app/src/main/res/layout/fragment_weapon_view_details_layout.xml
